+++ src/com/intellij/idea/plugin/hybris/common/utils/HybrisIcons.java
@@ -70,6 +70,7 @@ public final class HybrisIcons {
     public static final Icon EXTENSION_EXT = getIcon("/icons/extension/ext.svg");
     public static final Icon EXTENSION_OOTB = getIcon("/icons/extension/ootb.svg");
     public static final Icon EXTENSION_WEB = AllIcons.Nodes.WebFolder;
+    public static final Icon EXTENSION_COMMON_WEB = AllIcons.Nodes.WebFolder;
     public static final Icon EXTENSION_ADDON = AllIcons.Nodes.WebFolder;
     public static final Icon EXTENSION_BACKOFFICE = HYBRIS;
     public static final Icon EXTENSION_HMC = HYBRIS;