+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [2023.2.2]
+
+
+### Fixes
+- Add case-insensitive attribute suggestions for FlexibleSearch language [#484](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/484)
+
 ## [2023.2.1]
 
 ### Features